refactor: 重构数值计算系统，改用double作为默认数值类型
- 将整数字面量解析改为返回double类型
- 移除整数溢出检查，统一使用double进行数值计算
- 重构类型系统，统一使用double作为通用数值类型
- 更新所有相关测试用例以适配新的类型行为
- 重构内置函数，简化函数实现并统一返回double类型
- 为按位运算添加整数操作数校验，禁止布尔值参与位运算

+++ MathEval.Tests/ContextTests.cs
@@ -31,8 +31,8 @@ public class ContextTests {
     [Fact]
     public void SetFunction_WeakTyped_Works() {
         var ctx = new ExpressionContext();
-        ctx.SetFunction("add1", args => (long)args[0] + 1);
-        Assert.Equal(43L, Expression.Eval<long>("add1(42)", ctx));
+        ctx.SetFunction("add1", args => (double)args[0] + 1);
+        Assert.Equal(43.0, Expression.Eval<double>("add1(42)", ctx));
     }
 
     [Fact]
